requirements.txt
python-telegram-bot==13.17
requests
pandas
opencv-python
python-dotenv